+++ TorrentKit/STTorrentManager.mm
@@ -43,7 +43,7 @@
 @interface STFileEntry ()
 @property (readwrite, strong, nonatomic) NSString *name;
 @property (readwrite, strong, nonatomic) NSString *path;
-@property (readwrite, nonatomic) NSUInteger size;
+@property (readwrite, nonatomic) uint64_t size;
 @end
 
 #pragma mark -
